Fix horizontal scroll sync cast

+++ lib/view/asset_verification/details_group_page.dart
@@ -360,7 +360,6 @@ class _GroupAssetCardState extends State<_GroupAssetCard> {
                                 },
                                 icon: Icon(
                                   _isSignatureExpanded ? Icons.expand_less : Icons.expand_more,
-
                                 ),
                               ),
                             ],
@@ -516,7 +515,8 @@ class _GroupAssetCardState extends State<_GroupAssetCard> {
     }
     _isSyncingHorizontalScroll = true;
     final targetMax = target.position.maxScrollExtent;
-    final clampedOffset = source.offset.clamp(0, targetMax);
+    final clampedOffset = source.offset.clamp(0.0, targetMax).toDouble();
+
     if ((target.offset - clampedOffset).abs() > 0.5) {
       target.jumpTo(clampedOffset);
     }
